SNAKE-42 Fix issue with food not despawning
Note that this fix actually just ensures that only a single food instance can spawn per tile.  If it is later decided that multiple food instances can spawn on a tile, the StoryFoodSpawnTracker class's "occupiesPosition" method would need to be updated to detect all food instances on the tile, rather than just the first one it comes across.