SNAKE-65 Added support for showMap cutscene events
Note that the map had to be rendered to an sf::RenderTexture first, then turned into a sprite to be rendered to the screen.  This allows it to be faded in/out seamlessly with color or texture events.

For some reason, had to flip the rendered map texture vertically when rendering it to the screen...?

Not happy about the fact that the StoryMap objects are being created / destroyed every frame...